backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic[email]==2.9.2
email-validator==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12

